fix(dev): consider pages' route priority when matching request to a page
